requirements.txt
openai
google-generativeai
python-dotenv
pydantic
tqdm
